+++ Examples/MAX78000/Stepper_Control/motor_funcs.h
@@ -1,6 +1,6 @@
 
 
 void Debug_Motors(void);
-void rotate_revs(int slave_addr, double rotations);
+void rotate_revs(int slave_addr, float rotations);
 void target_tics(int slave_addr, int enc_tics);
 int Motor_Init_Settings();
